test: move to use a JavaString Action template
