+++ .gitattributes
@@ -1,0 +1,1 @@
+dist/** -diff linguist-generated=true
