+++ .github/codeql/codeql-config.yml
@@ -1,0 +1,5 @@
+name: "javascript-action CodeQL config"
+
+paths-ignore:
+  - node_modules
+  - dist
