+++ package.json
@@ -2,21 +2,25 @@
   "name": "junit2otel-collector",
   "version": "1.0.0",
   "description": "Collect GitHub Actions JUnit resutls and sent them to an OpenTelemetry service",
-  "main": "script.js",
+  "main": "main.js",
   "scripts": {
-    "test": "jest --ci --testResultsProcessor='jest-junit'"
+    "lint": "eslint .",
+    "prepare": "ncc build index.js -o dist --source-map --license licenses.txt",
+    "test": "jest --ci --testResultsProcessor='jest-junit'",
+    "all": "npm run lint && npm run prepare && npm run test"
   },
   "repository": {
     "type": "git",
     "url": "git+https://github.com/ivmavi/junit2otel-collector.git"
   },
   "keywords": [
+    "GitHub",
+    "Actions",
+    "JavaScript",
     "junit",
     "otel",
     "opentelemetry",
-    "test",
-    "github",
-    "actions"
+    "test"
   ],
   "author": "Kuisathaverat",
   "license": "Apache-2.0",
@@ -31,6 +35,8 @@
   "devDependencies": {
     "@actions/github": "^5.1.1",
     "jest": "^29.3.1",
-    "jest-junit": "^14.0.1"
+    "jest-junit": "^14.0.1",
+    "@vercel/ncc": "^0.31.1",
+    "eslint": "^8.0.0"
   }
 }
