+++ script.test.js
@@ -1,16 +1,15 @@
 const github = require("@actions/github");
 const script = require("./script");
 const core = require("@actions/core");
-const debug = core.debug;
 
 test("downloadReport", async () => {
   const { GITHUB_TOKEN } = process.env;
 
   const octokit = github.getOctokit(GITHUB_TOKEN);
-  owner = "ivmavi";
-  repo = "junit2otel-collector";
-  workflow_id = ".github/workflows/test.yml";
-  reportName = "test-results.xml";
+  var owner = "ivmavi";
+  var repo = "junit2otel-collector";
+  var workflow_id = ".github/workflows/test.yml";
+  var reportName = "test-results.xml";
 
   const runs = await octokit.rest.actions.listWorkflowRuns({
     owner,
@@ -18,9 +17,9 @@ test("downloadReport", async () => {
     workflow_id,
   });
 
-  workflowRunId = runs.data.workflow_runs[0].id;
+  var workflowRunId = runs.data.workflow_runs[0].id;
 
-  report = {
+  var report = {
     github: octokit,
     core: core,
     owner,
